+++ src/main/kotlin/lotto/provider/ticket/AutoTicketProvider.kt
@@ -1,6 +1,5 @@
 package lotto.provider.ticket
 
-import lotto.domain.AutoTicketProvideStrategy
 import lotto.domain.LottoTickets
 
 object AutoTicketProvider : LottoTicketsProvider(AutoTicketProvideStrategy) {
